+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# UNRELEASED
+
+-   Add new properties `roomKeyRequestsEnabled` and `roomKeyForwardingEnabled`
+    to OlmMachine.
+    ([#60](https://github.com/matrix-org/matrix-rust-sdk-crypto-wasm/pull/60),
+    ([matrix-rust-sdk#2902](https://github.com/matrix-org/matrix-rust-sdk/pull/2902))
+
 # matrix-sdk-crypto-wasm v3.2.0
 
 -   Add `timeout_secs` parameters to `OlmMachine.get_user_devices` and